Protect person record from endpoint
I think you missed one @malted!
