+++ src/app/harbor/battles/battles.tsx
@@ -21,7 +21,7 @@ import Cursed from './cursed'
 import pluralize from '../../../../lib/pluralize'
 import ProjectCard from './project-card'
 import { markdownComponents } from './mdc'
-import { Ship } from '@/app/utils/data'
+import { Ship } from '@/app/utils/server/data'
 import Modal from '@/components/ui/modal'
 import { sendFraudReport } from './fraud-utils'
 import { Button } from '@/components/ui/button'
