+++ src/app/harbor/shipyard/ship-utils.ts
@@ -2,10 +2,10 @@
 
 import { getSelfPerson } from '@/app/utils/server/airtable'
 import { getSession } from '@/app/utils/auth'
-import { fetchShips, person } from '@/app/utils/data'
+import { fetchShips, person } from '@/app/utils/server/data'
 import { getWakaSessions } from '@/app/utils/waka'
 import { cookies } from 'next/headers'
-import type { Ship } from '@/app/utils/data'
+import type { Ship } from '@/app/utils/server/data'
 import Airtable from 'airtable'
 import { withLock } from '../../../../lib/redis-lock'
 import sum from '../../../../lib/sum'
