+++ src/app/harbor/shipyard/ships.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useRef, useState } from 'react'
 import { motion, AnimatePresence } from 'framer-motion'
 import { Card } from '@/components/ui/card'
 import { stagedToShipped } from './ship-utils'
-import type { Ship } from '@/app/utils/data'
+import type { Ship } from '@/app/utils/server/data'
 import Image from 'next/image'
 import Icon from '@hackclub/icons'
 import ReactMarkdown from 'react-markdown'
@@ -21,7 +21,7 @@ import ReadmeHelperImg from '/public/readme-helper.png'
 import NewUpdateForm from './new-update-form'
 import Modal from '../../../components/ui/modal'
 import RepoLink from '@/components/ui/repo_link'
-import { EditableShipFields } from '../../utils/data'
+import { EditableShipFields } from '../../utils/server/data'
 import ThinkingDino from '/public/thinking.png'
 
 export default function Ships({
